Task 2: debug.c floats.

+++ zad2/debug.c
@@ -3,7 +3,7 @@
 
 void debug_matrix(const float *matrix, int size) {
   for (int i = 0; i < size; i++) {
-    printf("%d ", (int) matrix[i]);
+    printf("%f ", matrix[i]);
   }
   printf("\n");
 }
